+++ R/strcut.R
@@ -67,7 +67,7 @@
 #'
 #'
 #'
-#' @seealso [tinycodet_strings()]
+#' @seealso \link{tinycodet_strings}
 #'
 #'
 #' @examples